communications-micro/communications-micro/communication/requirements.txt
# Core Django packages
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Filtering and searching
django-filter>=23.0
django-extensions>=3.2.0

# File handling
Pillow>=10.0.0  # Image processing

# Authentication and permissions
djangorestframework-simplejwt>=5.2.0

# API documentation
drf-yasg>=1.21.0  # Swagger/OpenAPI docs

# Background tasks (optional)
celery>=5.3.0
redis>=4.5.0

# Email (if needed)
django-anymail>=10.0  # Email service integration

# Utilities
python-decouple>=3.8  # Environment variables
djangorestframework-recursive>=0.1.2

# Development and testing
pytest>=7.0.0
pytest-django>=4.5.0
factory-boy>=3.2.0
coverage>=7.0.0

# Production server
gunicorn>=21.0.0
whitenoise>=6.5.0  # Static file serving

# Monitoring and logging
sentry-sdk>=1.32.0

# Optional: If using advanced features
django-storages>=1.13.0  # Cloud storage (AWS S3, etc.)
boto3>=1.28.0  # AWS SDK (if using S3)